% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_energies.R
\name{combine_energies}
\alias{combine_energies}
\title{Reads in all st2 files and combines them into one aggregate data frame.}
\usage{
combine_energies(files)
}
\arguments{
\item{files}{A character vector contatining the full paths to the ST2 files per ligand.}
}
\value{
data frame of per residue interaction energy
}
\description{
Ensure each st2 file contatins the same number of rows OR the multiple simulations have the same number of frames.
}
