% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_relations.R
\name{energy_relations}
\alias{energy_relations}
\title{Calculates the point at which the ligand has the strongest ineteraction energy with the protein}
\usage{
energy_relations(ene_data, files, frames = 1250, lignodes = 50,
  energy_threshold = 100)
}
\arguments{
\item{ene_data}{data frame of combined interaction energy. derived from combine_energies() function.}

\item{files}{character vector contatining the full paths to the st2 files, identical to the one used in combine_energies() function.}

\item{frames}{number of frames in the input simulations for energy calculations (or rows containted in the .st2 files).}

\item{lignodes}{number of requested ligand nodes for the lig-path.}

\item{energy_threshold}{The energy threshold to included residues in the network.}
}
\value{
data frame contatining the From, TO and Energy columns which an igraph object can be created from
}
\description{
Calculates the point at which the ligand has the strongest ineteraction energy with the protein
}
