% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lom_matrix.R
\name{lom_matrix}
\alias{lom_matrix}
\title{Generates a matrix of count values which is used to create the ligand occupancy map}
\usage{
lom_matrix(directory, pmat, spacing = 75)
}
\arguments{
\item{directory}{Path containing the input .xyz file(s)}

\item{pmat}{A 4 x 4 Viewing Matrix used to transform the 3D coordinates to 2D. Callend by readCOM() function}

\item{spacing}{the size of the returend matrix ( 75 * 75 is the default)}
}
\value{
Matrix of count values used for heatmap generation

Example

pmat <- matrix(rep(0, 16), ncol = 4)
diag(pmat) <- 1
lom <- lom_matrix(directory = "~/Documents/xyz_files/", pmat = view_mat, spacing = 75)

plot_lom(lom)
}
\description{
see ?readCOM() to understand formatting of .xyz file.
}
