% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_res_to_centroid.R
\name{pdb_res_to_centroid}
\alias{pdb_res_to_centroid}
\title{Reads in a PDB file and converts atomic coordinates to residue centroids XYZ.}
\usage{
pdb_res_to_centroid(pdb = "input.pdb")
}
\arguments{
\item{pdb}{protein data bank file}
}
\value{
data frame of per residue centroids containing Residue Number, Residue Name, x, y and z coordinates.
}
\description{
Reads in a PDB file and converts atomic coordinates to residue centroids XYZ.
}
