% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lom.R
\name{plot_lom}
\alias{plot_lom}
\title{Plots the ligand occupancy map.}
\usage{
plot_lom(lom)
}
\arguments{
\item{lom}{the ligand occupancy map matrix derived using the lom_matrix() function}
}
\value{
NULL; Plots lom as a heatmap
}
\description{
The map uses a 6 gradient colour scale with the following colours:
(low Count) "darkblue", "cyan", "green", "yellow", "orange", "red" (High Count)
}
