% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_st2.R
\name{read_st2}
\alias{read_st2}
\title{Reads in st2 files in a given directory}
\usage{
read_st2(st2dir, as.sum = FALSE)
}
\arguments{
\item{st2dir}{single directory contatining all st2 files}

\item{as.sum}{Default is FALSE which seperates VdW and Coloumb Energies}
}
\value{
List of interaction energies.
}
\description{
Reads in st2 files in a given directory
}
